# api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AudioBookRequest",
    "description": "Your tool for handling audiobook requests on a Plex/Audiobookshelf/Jellyfin instance",
    "version": "1.3.0",
    "contact": {
      "name": "AudioBookRequest",
      "url": "https://github.com/markbeep/AudioBookRequest"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/markbeep/AudioBookRequest/blob/main/LICENSE"
    }
  },
  "paths": {
...
    "/api/v1/status": {
      "get": {
        "tags": [
          "API",
          "System"
        ],
        "summary": "Get Status",
        "description": "Get detailed system status information.\n\n**Requires:** Admin privileges\n\n**Authentication:** Bearer token (API key)\n\nReturns comprehensive system status including database health,\nuser counts, request counts, and version information.",
        "operationId": "get_status_api_v1_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "API",
          "Users"
        ],
        "summary": "List Users",
        "description": "List all users in the system.\n\n**Requires:** Admin privileges\n\n**Authentication:** Bearer token (API key)\n\nReturns a paginated list of all users with their basic information.",
        "operationId": "list_users_api_v1_users_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of users to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of users to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of users to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of users to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "API",
          "Users"
        ],
        "summary": "Create New User",
        "description": "Create a new user in the system.\n\n**Requires:** Admin privileges\n\n**Authentication:** Bearer token (API key)\n\nCreates a new user with the specified username, password, and group.\nUsername must be unique within the system.",
        "operationId": "create_new_user_api_v1_users_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{username}": {
      "get": {
        "tags": [
          "API",
          "Users"
        ],
        "summary": "Get User",
        "description": "Get information about a specific user.\n\n**Requires:** Admin privileges\n\n**Authentication:** Bearer token (API key)\n\nReturns detailed information about the specified user.",
        "operationId": "get_user_api_v1_users__username__get",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "API",
          "Users"
        ],
        "summary": "Update User",
        "description": "Update an existing user.\n\n**Requires:** Admin privileges\n\n**Authentication:** Bearer token (API key)\n\nUpdates the specified user's password and/or group.\nRoot users cannot have their group changed.",
        "operationId": "update_user_api_v1_users__username__put",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "API",
          "Users"
        ],
        "summary": "Delete User",
        "description": "Delete a user from the system.\n\n**Requires:** Admin privileges\n\n**Authentication:** Bearer token (API key)\n\nPermanently removes the specified user from the system.\nCannot delete own user or root users.",
        "operationId": "delete_user_api_v1_users__username__delete",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "API",
          "Users"
        ],
        "summary": "Get Current User",
        "description": "Get current user's information.\n\n**Requires:** Any authenticated user\n\n**Authentication:** Bearer token (API key)\n\nReturns information about the user associated with the provided API key.",
        "operationId": "get_current_user_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GroupEnum": {
        "type": "string",
        "enum": [
          "untrusted",
          "trusted",
          "admin"
        ],
        "title": "GroupEnum"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "System status"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Application version"
          },
          "database": {
            "type": "string",
            "title": "Database",
            "description": "Database status"
          },
          "users_count": {
            "type": "integer",
            "title": "Users Count",
            "description": "Total number of users"
          },
          "requests_count": {
            "type": "integer",
            "title": "Requests Count",
            "description": "Total number of book requests"
          },
          "api_keys_count": {
            "type": "integer",
            "title": "Api Keys Count",
            "description": "Total number of active API keys"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Current timestamp"
          }
        },
        "type": "object",
        "required": [
          "status",
          "version",
          "database",
          "users_count",
          "requests_count",
          "api_keys_count",
          "timestamp"
        ],
        "title": "StatusResponse",
        "description": "System status response model"
      },
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Username",
            "description": "Unique username"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "description": "User password"
          },
          "group": {
            "$ref": "#/components/schemas/GroupEnum",
            "description": "User group (untrusted, trusted, admin)",
            "default": "untrusted"
          },
          "root": {
            "type": "boolean",
            "title": "Root",
            "description": "Whether to create as root admin user",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserCreate",
        "description": "User creation request model"
      },
      "UserResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Unique username"
          },
          "group": {
            "$ref": "#/components/schemas/GroupEnum",
            "description": "User group determining permissions"
          },
          "root": {
            "type": "boolean",
            "title": "Root",
            "description": "Whether this is the root admin user"
          }
        },
        "type": "object",
        "required": [
          "username",
          "group",
          "root"
        ],
        "title": "UserResponse",
        "description": "User information response model"
      },
      "UserUpdate": {
        "properties": {
          "password": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "New password (optional)"
          },
          "group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupEnum"
              },
              {
                "type": "null"
              }
            ],
            "description": "New user group (optional)"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "User update request model"
      },
      "UsersListResponse": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "type": "array",
            "title": "Users",
            "description": "List of users"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of users"
          }
        },
        "type": "object",
        "required": [
          "users",
          "total"
        ],
        "title": "UsersListResponse",
        "description": "Users list response model"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key authentication using Bearer token"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}